# gen/http/master/server/encode_decode.go
// Code generated by goa v3.23.1, DO NOT EDIT.
//
// master HTTP server encoders and decoders
//
// Command:
// $ goa gen stock-bot/design

package server

import (
	"context"
	"net/http"
	master "stock-bot/gen/master"
	masterviews "stock-bot/gen/master/views"

	goahttp "goa.design/goa/v3/http"
)

// EncodeGetStockResponse returns an encoder for responses returned by the
// master get_stock endpoint.
func EncodeGetStockResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res := v.(*masterviews.StockbotStockMaster)
		enc := encoder(ctx, w)
		body := NewGetStockResponseBody(res.Projected)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeGetStockRequest returns a decoder for requests sent to the master
// get_stock endpoint.
func DecodeGetStockRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*master.GetStockPayload, error) {
	return func(r *http.Request) (*master.GetStockPayload, error) {
		var (
			symbol string

			params = mux.Vars(r)
		)
		symbol = params["symbol"]
		payload := NewGetStockPayload(symbol)

		return payload, nil
	}
}