# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Stock Bot Service
    description: Service for placing and managing stock orders
    version: 0.0.1
host: localhost:8080
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
...
    /master/stocks/{symbol}:
        get:
            tags:
                - master
            summary: get_stock master
            description: Get basic master data for a single stock.
            operationId: master#get_stock
            parameters:
                - name: symbol
                  in: path
                  description: Stock symbol to look up
                  required: true
                  type: string
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/StockbotStockMaster'
            schemes:
                - http
...
definitions:
...
    StockbotStockMaster:
        title: 'Mediatype identifier: application/vnd.stockbot.stock-master; view=default'
        type: object
        properties:
            industry_code:
                type: string
                description: 業種コード
                example: Eligendi et consequatur maxime maxime porro nam.
            industry_name:
                type: string
                description: 業種コード名
                example: Aliquid provident dolores laboriosam eaque quas aut.
            market:
                type: string
                description: 優先市場
                example: Pariatur suscipit corrupti.
            name:
                type: string
                description: 銘柄名
                example: Voluptatem quas sunt magnam et.
            name_kana:
                type: string
                description: 銘柄名（カナ）
                example: In perferendis quia.
            symbol:
                type: string
                description: 銘柄コード
                example: Illo vel nemo.
        description: get_stock_response_body result type (default view)
        example:
            industry_code: Earum perspiciatis et non.
            industry_name: Rem vitae numquam blanditiis.
            market: Qui molestiae.
            name: Eos quisquam.
            name_kana: Aut commodi sunt nobis maiores veritatis.
            symbol: Repudiandae ab laudantium itaque.
        required:
            - symbol
            - name
            - market